refactor(applications): standardize validator namespace
- Rename Applications.Validations to Applications.Validators for consistency
- Update namespace in CourseValidator, PersonValidator, ProfessorValidator, ProgramValidator, and StudentValidator
